requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
pandas==2.1.4
python-dotenv==1.0.0
google-generativeai==0.3.2
requests==2.31.0
Pillow==10.1.0
openpyxl==3.1.2
